Update Inventory.java
add a new method "readshelf" to get the shelf number and its remaining capacity, put into a map. When add a new item, it will find if the Qty< the largest remaining capacity. If true, the new-added item will go to that shelf;   if false, it will go to a new(empty) shelf.